+++ src/extension.ts
@@ -115,6 +115,9 @@ export function activate(context: vscode.ExtensionContext): void {
     vscode.commands.registerCommand("rojoExplorer.createRemoteFunction", (node?: ExplorerNode) =>
       createResource(provider, fileSystem, node, "RemoteFunction"),
     ),
+    vscode.commands.registerCommand("rojoExplorer.createBindableEvent", (node?: ExplorerNode) =>
+      createResource(provider, fileSystem, node, "BindableEvent"),
+    ),
     vscode.commands.registerCommand("rojoExplorer.createStringValue", (node?: ExplorerNode) =>
       createResource(provider, fileSystem, node, "StringValue"),
     ),
@@ -785,6 +788,8 @@ function localizeResourceKind(kind: CreatableResourceKind): string {
       return vscode.l10n.t("RemoteEvent");
     case "RemoteFunction":
       return vscode.l10n.t("RemoteFunction");
+    case "BindableEvent":
+      return vscode.l10n.t("BindableEvent");
     case "StringValue":
       return vscode.l10n.t("StringValue");
     case "LocalizationTable":